feat(HTD-52):  refactor masonry grid & fetch posts to make correct suspense
