+++ src/components/sections/blog/overview-content/overview-content.styles.ts
@@ -25,13 +25,3 @@ export const StyledSubscriberBox = styled.aside`
     min-width: 387px;
   }
 `;
-
-export const StyledLoadMoreWrapper = styled.div`
-  justify-content: center;
-
-  @media (min-width: ${screens.lg}) {
-    display: flex;
-    justify-content: flex-start;
-    padding-left: 151px;
-  }
-`;
